# apigateway/docs/swagger.yaml
basePath: /
definitions:
  proto.Movie:
    properties:
      id:
        type: integer
      title:
        type: string
      year:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Gateway for Movies Service
  title: API Gateway Swagger
  version: "1.0"
paths:
  /movies:
    get:
      consumes:
      - application/json
      description: Retorna uma lista paginada de filmes cadastrados
      parameters:
      - description: Número da página (padrão 1)
        in: query
        name: pageNumber
        type: integer
      - description: Resultados por página (padrão 10)
        in: query
        name: resultsPerPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de filmes retornada com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parâmetro inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Página não encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Listar filmes
      tags:
      - Movies
    post:
      consumes:
      - application/json
      description: Registra um novo filme no banco de dados
      parameters:
      - description: Objeto do filme a ser criado
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/proto.Movie'
      produces:
      - application/json
      responses:
        "201":
          description: Filme criado com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Requisição inválida
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Criar novo filme
      tags:
      - Movies
  /movies/{id}:
    delete:
      consumes:
      - application/json
      description: Remove um filme do banco de dados
      parameters:
      - description: ID do filme
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Filme deletado com sucesso
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Filme não encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Deletar filme por ID
      tags:
      - Movies
    get:
      consumes:
      - application/json
      description: Retorna um único filme com base em seu ID
      parameters:
      - description: ID do filme
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Filme retornado com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Filme não encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Obter filme por ID
      tags:
      - Movies
schemes:
- http
swagger: "2.0"